requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
streamlit
python-dotenv
tiktoken